Update `actions/checkout` in GitHub Actions workflows to 3.x
A new major version series of the `actions/checkout` GitHub Action action was started in March 2022. This project's
GitHub Actions workflows previously used the `v2` ref when specifying the version of the action to use, resulting in the
workflows not having access to the advancements made to the action in the 3.x series.

One of these advancements is the update from using Node.js 12.x to Node.js 16.x for the action on the GitHub Actions
runner. GitHub has deprecated the use of Node.js 12.x in actions, which results in warnings being shown in the workflow
run summary page of every GitHub actions workflow that uses a Node.js 12.x-based action.

Updating the action resolves the warnings. I have carefully evaluated the changes which caused the major version bump in
the action and found that they do not have any impact on the functionality of this project's workflows.